+++ incl/posts.nim
@@ -1,0 +1,3 @@
+
+const
+  externalIcon* = "<sup style=\"font-size:0.5em\">↗</sup>"