+++ index.nim
@@ -12,6 +12,7 @@ nbText "# thisago's blog"
 nbText "## Posts"
 
 for post in getAllposts postsDir:
-  nbText fmt"- [{post.toPostName}]({post}.html)"
+  nbText fmt"### [{post.toPostName}]({post}.html)"
+  # TODO: Add creation and modification date for each post
 
 nbSave